Remove createGoal() and detect goal in player logic instead
The only difference between a regular wall and the goal was the 'goal' tag. Also, the goal object was triggering a level change on *every collision* with anything. That logic is better suited in the player object, since that should only be triggered when *the player collides with something that is a goal*.