Implemented multiple levels and automatic level progression
The app keeps track of which level index it's currently at and will increment that once the nextLevel() function is called. If trying to load a level at an index bigger than the number of levels the counter will reset and no level is loaded. Whichever scene is at the top of the stack will be running. There's also a lives counter, and when the playerDied() function is called the lives counter decreases and the level reloads.