feat: added pickup item
The item removes itself when colliding with an entity with the 'player' tag. The Entity class now sends a 'destroy' message to all its components. PhysicsComponent react to this message by removing its physics Body from the physics world and then deleting the body. The PhysicsComponent is now prepared to attach any event to the Body. Right now Body only cares to call onCollision.